# specs/003-add-supabase-and-payment/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Podscript Auth & Payment API
  description: API endpoints for user authentication, credits management, and payment integration
  version: 1.0.0
  contact:
    name: Podscript
servers:
  - url: http://localhost:8001
    description: Development server
  - url: https://your-domain.com
    description: Production server
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Create a new account with email and password. Grants 10 free credits.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Sets access_token cookie
        '400':
          description: Invalid email or password format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login existing user
      description: Authenticate with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Sets access_token cookie
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout current user
      description: Clear session and cookies
      operationId: logout
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Clears access_token cookie
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user info
      description: Returns current user's email and credit balance
      operationId: getCurrentUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User info retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============ CREDITS ENDPOINTS ============
  /api/credits/balance:
    get:
      tags: [Credits]
      summary: Get credit balance
      description: Returns current user's credit balance
      operationId: getCreditBalance
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Balance retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/credits/transactions:
    get:
      tags: [Credits]
      summary: Get transaction history
      description: Returns paginated list of credit transactions
      operationId: getTransactionHistory
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Number of transactions to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Offset for pagination
      responses:
        '200':
          description: Transactions retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionHistoryResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============ PAYMENT ENDPOINTS ============
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT token stored in HTTPOnly cookie
  schemas:
...
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 6
          example: "password123"
    AuthResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        credit_balance:
          type: integer
          example: 10
    UserInfoResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        credit_balance:
          type: integer

    # ---- Credits Schemas ----
    BalanceResponse:
      type: object
      properties:
        balance:
          type: integer
          example: 50
    TransactionHistoryResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/CreditTransaction'
        total:
          type: integer
          description: Total number of transactions
    CreditTransaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [recharge, consumption, bonus, refund]
        amount:
          type: integer
          description: Positive for additions, negative for deductions
        balance_after:
          type: integer
        description:
          type: string
          example: "+50 充值"
        created_at:
          type: string
          format: date-time

    # ---- Payment Schemas ----
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Not authenticated"
...